+++ Detector/DetectorPoint.cxx
@@ -14,9 +14,7 @@ SHiP::DetectorPoint::DetectorPoint(Int_t eventID, Int_t trackID, Int_t detID,
     : FairMCPoint(trackID, detID, pos, mom, tof, length, eLoss, eventID),
       fPdgCode(pdgcode),
       fLpos{Lpos.X(), Lpos.Y(), Lpos.Z()},
-      fLmom{Lmom.X(), Lmom.Y(), Lmom.Z()} {
-  setDetectorName();
-}
+      fLmom{Lmom.X(), Lmom.Y(), Lmom.Z()} {}
 
 // -----   Standard constructor  ----------------
 SHiP::DetectorPoint::DetectorPoint(Int_t eventID, Int_t trackID, Int_t detID,
@@ -26,13 +24,11 @@ SHiP::DetectorPoint::DetectorPoint(Int_t eventID, Int_t trackID, Int_t detID,
     : FairMCPoint(trackID, detID, pos, mom, tof, length, eLoss, eventID),
       fPdgCode(pdgcode),
       fLpos{pos.X(), pos.Y(), pos.Z()},
-      fLmom{mom.X(), mom.Y(), mom.Z()} {
-  setDetectorName();
-}
+      fLmom{mom.X(), mom.Y(), mom.Z()} {}
 
 // -----   Public method Print   -------------------------------------------
 void SHiP::DetectorPoint::Print() const {
-  LOG(info) << "-I- " << fDetectorName << " point for track " << fTrackID
+  LOG(info) << "-I- " << DetectorName() << " point for track " << fTrackID
             << " in detector " << fDetectorID;
   LOG(info) << "    Position (" << fX << ", " << fY << ", " << fZ << ") cm";
   LOG(info) << "    Momentum (" << fPx << ", " << fPy << ", " << fPz << ") GeV";
@@ -41,10 +37,6 @@ void SHiP::DetectorPoint::Print() const {
   extraPrintInfo();
 }
 // -------------------------------------------------------------------------
-
-void SHiP::DetectorPoint::setDetectorName() {
-  fDetectorName = "GenericDetector";
-}
 
 void SHiP::DetectorPoint::extraPrintInfo() const {
   LOG(info) << "Nothing to see here";