+++ Detector/DetectorPoint.h
@@ -35,14 +35,13 @@ class DetectorPoint : public FairMCPoint {
   TVector3 LastPoint() const { return TVector3(fLpos[0], fLpos[1], fLpos[2]); }
   TVector3 LastMom() const { return TVector3(fLmom[0], fLmom[1], fLmom[2]); }
 
-  virtual void setDetectorName();
+  virtual const char* DetectorName() const { return "GenericDetector"; }
 
   virtual void extraPrintInfo() const;
 
  protected:
   Int_t fPdgCode;
   std::array<Double_t, 3> fLpos, fLmom;
-  std::string fDetectorName = "detector";
 
  private:
   ClassDef(SHiP::DetectorPoint, 1);