+++ UpstreamTagger/UpstreamTaggerPoint.cxx
@@ -16,10 +16,6 @@ UpstreamTaggerPoint::UpstreamTaggerPoint() : SHiP::DetectorPoint() {}
 UpstreamTaggerPoint::~UpstreamTaggerPoint() {}
 // -------------------------------------------------------------------------
 
-void UpstreamTaggerPoint::setDetectorName() {
-  fDetectorName = "UpstreamTagger";
-}
-
 void UpstreamTaggerPoint::extraPrintInfo() const {
   LOG(info) << "This is the upstream tagger";
 }