+++ UpstreamTagger/UpstreamTaggerPoint.h
@@ -22,7 +22,8 @@ class UpstreamTaggerPoint : public SHiP::DetectorPoint {
   /** Destructor **/
   virtual ~UpstreamTaggerPoint();
 
-  void setDetectorName() override;
+  const char* DetectorName() const override { return "UpstreamTagger"; }
+
   void extraPrintInfo() const override;
 
  private: