+++ strawtubes/strawtubesPoint.h
@@ -37,7 +37,7 @@ class strawtubesPoint : public SHiP::DetectorPoint {
 
   Double_t dist2Wire() const { return fdist2Wire; }
 
-  void setDetectorName() override;
+  const char* DetectorName() const override { return "strawtubesDetector"; }
   void extraPrintInfo() const override;
 
  private: