+++ veto/vetoPoint.h
@@ -33,7 +33,7 @@ class vetoPoint : public SHiP::DetectorPoint {
   /** Destructor **/
   virtual ~vetoPoint();
 
-  void setDetectorName() override;
+  const char* DetectorName() const override { return "vetoDetector"; }
   void extraPrintInfo() const override;
 
   /** Output to screen **/